+++ example/AppQuery.ts
