+++ example/svelte.config.js
@@ -1,0 +1,8 @@
+const sveltePreprocess = require('svelte-preprocess');
+
+module.exports = {
+	preprocess: sveltePreprocess({
+		// ...svelte-preprocess options
+	}),
+	// ...other svelte options
+};
