+++ example/tsconfig.json
@@ -1,0 +1,8 @@
+{
+	"extends": "../tsconfig.json",
+	"compilerOptions": {
+		"strict": true,
+		"strictNullChecks": false,
+		"lib": ["dom", "esnext"],
+	}
+}
