+++ package.json
@@ -22,7 +22,9 @@
     "parcel-plugin-svelte": "^4.0.6",
     "prettier": "^2.0.5",
     "relay-compiler": "^9.1.0",
+    "relay-compiler-language-typescript": "^12.0.2",
     "relay-config": "^9.1.0",
-    "typescript": "^3.9.2"
+    "svelte-preprocess": "^3.7.4",
+    "typescript": "^3.9.3"
   }
 }
