+++ relay.config.js
@@ -3,4 +3,5 @@ module.exports = {
 	// Configuration options accepted by the `relay-compiler` command-line tool and `babel-plugin-relay`.
 	src: './example',
 	schema: './example/server/schema.graphql',
+	language: 'typescript'
 };
